Added TQueue, reworked some of the planetary physics

+++ ProjectBane/Bane-Common/Common/Hash.h
@@ -17,8 +17,9 @@ inline uint64 GetDJB264BitHash(const char* Ptr, size_t Length)
 
 constexpr uint64 GetDJB264BitHash_ConstExpr_Impl(const char* String, uint64 InValue, uint32 Index)
 {
-	return String[Index] != 0 ? GetDJB264BitHash_ConstExpr_Impl(String, (((InValue << 5) + InValue) + String[Index]), Index + 1) : InValue;
+	return String[Index] != 0 ? GetDJB264BitHash_ConstExpr_Impl(String, (((InValue << 5) + (uint64)InValue) + (uint64)String[Index]), Index + 1) : InValue;
 }
+
 
 constexpr uint64 GetDJB264BitHash_ConstExpr(const char* InString)
 {
